pyproject.toml
[tool.poetry]
name = "muninn"
version = "0.1.0"
description = "AI assistant bot"
authors = ["Divan Visagie <me@divanv.com>"]
license = "Private"
readme = "README.md"

[virtualenvs]
in-project = true

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
duckduckgo-search = "^2.8.5"
notion-client = "^2.0.0"
spacy = "^3.5.1"
python-telegram-bot = "^20.1"
pymongo = "^4.3.3"
transformers = "^4.27.0"
torch = "^2.0.0"
torchvision = "^0.15.1"
torchaudio = "^2.0.1"
mypy = "^1.1.1"
pytest-spec = "^3.2.0"
pytest-describe = "^2.0.1"
pytest-watch = "^4.2.0"
pykka = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
poethepoet = "^0.18.1"

[tool.poe.tasks]
integration = "pytest tests/integration"
unit = {shell = "python3 activate.py; pytest tests/unit --spec" }
model = "pytest tests/model"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
